* master: (229 commits)
  Remove :yaml related tests and fix other related to parsing empty arrays
  Fix warning: & interpreted as argument prefix
  Fix syntax error and remove duplicated test
  Add CHANGELOG entry for YAML parsing removal
  remove yaml as a param parser :burn:
  mv README README.rdoc as app template file
  improved different_target conditions
  CVE-2013-0156: Safe XML params parsing. Doesn't allow symbols or yaml.
  * Strip nils from collections on JSON and XML posts. [CVE-2013-0155] * dealing with empty hashes. Thanks Damien Mathieu
  Revert "Merge branch 'master-sec'"
  CVE-2013-0156: Safe XML params parsing. Doesn't allow symbols or yaml.
  view_cache_dependency API
  Revert "unpermitted params" exception -- it's just not going to work. See the discussion on https://github.com/rails/strong_parameters/pull/75.
  Never treat action or controller as unpermitted params
  Bump rack dependency to 1.4.3
  Eliminate Rack::File headers deprecation warning
  set 'alt' attribute for image_submit_tag
  Do not generate local vars for partials without object or collection
  Ignore binds payload with nil column in AR log subscriber
  Use absolute path
  ...

+++ install.rb
@@ -1,11 +1,16 @@
 version = ARGV.pop
 
+if version.nil?
+  puts "Usage: ruby install.rb version"
+  exit(64)
+end
+
 %w( activesupport activemodel activerecord actionpack actionmailer railties ).each do |framework|
   puts "Installing #{framework}..."
-  `cd #{framework} && gem build #{framework}.gemspec && gem install #{framework}-#{version}.gem --no-ri --no-rdoc && rm #{framework}-#{version}.gem`
+  `cd #{framework} && gem build #{framework}.gemspec && gem install #{framework}-#{version}.gem --local --no-ri --no-rdoc && rm #{framework}-#{version}.gem`
 end
 
 puts "Installing Rails..."
 `gem build rails.gemspec`
-`gem install rails-#{version}.gem --no-ri --no-rdoc `
+`gem install rails-#{version}.gem --local --no-ri --no-rdoc `
 `rm rails-#{version}.gem`
